requirements.txt
streamlit
pandas
google-genai
configparser